fixed temp dir, some doc

+++ ciciol/backends/notify.py
@@ -7,6 +7,7 @@ import tempfile
 import urllib
 import logging
 import os
+import shutil
 
 logger = logging.getLogger(__name__)
 
@@ -15,6 +16,12 @@ class NotifyBackend(object):
     """
     libnotify backend for Ciciol
     """
+    def __init__(self):
+        self.tempdir = tempfile.mkdtemp()
+
+    def __del__(self):
+        shutil.rmtree(self.tempdir)
+
     def notify(self, notifications):
         """
         notify method
@@ -26,14 +33,12 @@ class NotifyBackend(object):
             except ValueError:
                 sender, message = notification
             if photo is not None:
-                tempdir = tempfile.mkdtemp()
-                tmp = os.path.join(tempdir, photo.split("/")[-1])
-
-                logging.info("Getting image url %s to %s", photo, tmp)
-                photo_content = urllib.urlopen(photo)
-                with open(tmp, "w") as f:
-                    f.write(photo_content.read())
-                logging.info("Saved... %s", tmp)
+                tmp = os.path.join(self.tempdir, photo.split("/")[-1])
+                if not os.path.exists(tmp):
+                    logging.info("Saving image url %s to %s", photo, tmp)
+                    photo_content = urllib.urlopen(photo)
+                    with open(tmp, "w") as f:
+                        f.write(photo_content.read())
                 self._notify_one(sender, message, tmp)
             else:
                 self._notify_one(sender, message)
